Add tmux integration
This change requires you have `christoomey/vim-tmux-navigator` and
`termhn/i3-vim-nav` vim plugins installed. In the future it might be
best to unify these two plugins.
The binary now uses tmux directly to switch panes or send keys to the
vim process. This requires some additional mappings in vim to
distinguish between tmux and i3 when moving out of the vim window.